Refactor LoadTestBase
