+++ .github/workflows/beam_PostCommit_Java_IO_Performance_Tests.yml
@@ -64,7 +64,7 @@ jobs:
       matrix:
         job_name: ["beam_PostCommit_Java_IO_Performance_Tests"]
         job_phrase: ["Run Java PostCommit IO Performance Tests"]
-        test_case: ["GCSPerformanceTest", "BigTablePerformanceTest", "BigQueryStorageApiStreamingPerformanceTest"]
+        test_case: ["IcebergPerformanceTest"]
     steps:
     - uses: actions/checkout@v7
     - name: Setup repository
