+++ it/common/build.gradle
@@ -28,9 +28,16 @@ ext.summary = "Code used by all integration test utilities."
 dependencies {
     implementation enforcedPlatform(library.java.google_cloud_platform_libraries_bom)
     implementation project(path: ":sdks:java:core", configuration: "shadow")
+    implementation project(path: ":runners:google-cloud-dataflow-java")
+    implementation project(path: ":sdks:java:testing:test-utils")
+    implementation 'com.google.cloud:google-cloud-bigquery'
+    implementation 'com.google.cloud:google-cloud-monitoring'
+    provided 'com.google.api.grpc:proto-google-cloud-monitoring-v3'
+    implementation library.java.gax
     implementation library.java.google_api_services_dataflow
     implementation library.java.google_auth_library_credentials
     implementation library.java.google_auth_library_oauth2_http
+    implementation library.java.google_api_services_bigquery
     implementation library.java.vendored_guava_32_1_2_jre
     implementation library.java.slf4j_api
     implementation library.java.commons_lang3
@@ -38,6 +45,8 @@ dependencies {
     implementation library.java.google_code_gson
     implementation library.java.google_http_client
     implementation library.java.guava
+    implementation library.java.protobuf_java_util
+    implementation library.java.protobuf_java
 
     // TODO: excluding Guava until Truth updates it to >32.1.x
     testImplementation(library.java.truth) {
@@ -46,4 +55,5 @@ dependencies {
     testImplementation library.java.junit
     testImplementation library.java.mockito_inline
     testRuntimeOnly library.java.slf4j_simple
+    testImplementation project(path: ":sdks:java:extensions:protobuf", configuration: "testRuntimeMigration")
 }
