+++ it/common/src/main/java/org/apache/beam/it/common/bigquery/BigQueryResourceManager.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.it.gcp.bigquery;
+package org.apache.beam.it.common.bigquery;
 
 import com.google.api.gax.paging.Page;
 import com.google.auth.Credentials;
