+++ it/common/src/main/java/org/apache/beam/it/common/monitoring/package-info.java
@@ -15,12 +15,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.it.gcp.bigquery;
 
-/** Custom exception for {@link BigQueryResourceManager} implementations. */
-public class BigQueryResourceManagerException extends RuntimeException {
-
-  public BigQueryResourceManagerException(String errorMessage, Throwable err) {
-    super(errorMessage, err);
-  }
-}
+/** Package for querying metrics from cloud monitoring. */
+package org.apache.beam.it.common.monitoring;
