+++ it/google-cloud-platform/src/main/java/org/apache/beam/it/gcp/bigquery/conditions/BigQueryRowsCheck.java
@@ -20,7 +20,7 @@ package org.apache.beam.it.gcp.bigquery.conditions;
 import com.google.auto.value.AutoValue;
 import com.google.cloud.bigquery.TableId;
 import org.apache.beam.it.conditions.ConditionCheck;
-import org.apache.beam.it.gcp.bigquery.BigQueryResourceManager;
+import org.apache.beam.it.common.bigquery.BigQueryResourceManager;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 /** ConditionCheck to validate if BigQuery has received a certain number of rows. */
