+++ it/google-cloud-platform/src/main/java/org/apache/beam/it/gcp/dataflow/FlexTemplateClient.java
@@ -32,6 +32,8 @@ import com.google.auth.Credentials;
 import com.google.auth.http.HttpCredentialsAdapter;
 import dev.failsafe.Failsafe;
 import java.io.IOException;
+
+import org.apache.beam.it.common.dataflow.AbstractPipelineLauncher;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
