+++ it/iceberg/src/main/resources/test-artifact.json
@@ -1,0 +1,1 @@
+["This is a test artifact."]
