+++ it/iceberg/src/test/java/org/apache/beam/it/iceberg/IcebergIOLT.java
@@ -1,0 +1,7 @@
+package org.apache.beam.it.iceberg;
+
+import org.apache.beam.it.common.IOLoadTestBase;
+
+public class IcebergIOLT extends IOLoadTestBase {
+
+}
