+++ settings.gradle.kts
@@ -122,6 +122,7 @@ include(":it:conditions")
 include(":it:datadog")
 include(":it:elasticsearch")
 include(":it:google-cloud-platform")
+include(":it:iceberg")
 include(":it:jdbc")
 include(":it:kafka")
 include(":it:testcontainers")
